Isolation of udev code from remaining systemd
This commit is a first attempt to isolate the udev code from the
remaining code base.  It intentionally does not modify any files
but purely delete files which, on a first examination, appear to
not be needed.  This is a sweeping commit which may easily have
missed needed code.  Files can be retrieved by doing a checkout
from the previous commit:

 git checkout 2944f347d0 -- <filename>