Initial revamping of the build system
This is the first pass attempting to keep as much of the build system
as is necessary for only udev from the fork.  Emphasis was given to
configure.ac.  Gutting had to be done to Makefile.am but this needs
work to be broken out into SUBDIR Makefile.am which each address those
pieces.