Eliminate kmod dependency
Changes to rules were introduced by
06316d9f1a91b4d3efdb7402e72498cb3deb1806 while kmod support was
introduced in other commits. A ton of commits were made involving kmod
and it is quite clear that it is broken, so we remove it.

This changes our rules to depend on modprobe. As long as the modprobe
binary is in /, and not /usr, udev module loading should function
properly.

Signed-off-by: Richard Yao <ryao@gentoo.org>