Define unshare() macro to avoid dependence on __GNU_SOURCE
unshare() is a linux syscall that we need. Unfortunately, access to it
depends on __GNU_SOURCE, which we would like to avoid. As such, we
define a macro to invoke it by its syscall code. This is necessary to
avoid the following warning from Clang:

implicit declaration of function 'unshare' is invalid in C99
[-Wimplicit-function-declaration]

Signed-off-by: Richard Yao <ryao@gentoo.org>