test/Makefile.am: fix tests for make distcheck
At several points in the forked code, top_srcdir and top_builddir
are mixed up.  It is not clear if this is an upstream bug or if this
is a result of the initial restructuring of the build system.  This
commit addresses this issue for tests triggered during make distcheck.

Signed-off-by: Anthony G. Basile <blueness@gentoo.org>