configure.ac: remove unecessary check for python
The original checks for python were to build some man pages and
for systemd-python.  These checks are no longer relevant.  There
still is a dependency on python for tests but this is checked by
test/rules-test.sh which tests for python or skips.  Also, there
is a dependency on perl for both tests and hwdb.  These checks can
be reintroduced at a future date when re revisit both.

Signed-off-by: Anthony G. Basile <blueness@gentoo.org>