Replace unused DISTCLEAN_LOCAL_HOOKS and correct path in test/Makefile.am
The DISTCLEAN_LOCAL_HOOKS variable is not used by automake when generating
Makefiles.  This has been replaced with the clean-local target which does get
used.  The clean-local target is defined to be dependent on the
test-sys-distclean target as recommended in the automake manual.

The paths in test-sys-distclean have been corrected to ensure that test/sys is
deleted during 'make distcheck'.