Allow autogen on systems without gtk-doc or introspection installed
On systems without gtk-doc installed, we were mocking up docs/gtk-doc.make
and m4/gtk-doc.m4 to avoid a failure in configure.  On systems without
introspection installed, we were doing nothing and allowing the failure
to occur.  Since autogen should only be run on the developer side this is
should be okay, but to be more friendly to other distros and users that
want to start from autogen, we now bundle

	docs/gtk-doc.make
	m4/gtk-doc.m4
	m4/introspection.m4

Care in the future should be taken by the developers to make sure these
files stay up to day.

This should take care of

	https://github.com/gentoo/eudev/issues/11

Signed-off-by: Anthony G. Basile <blueness@gentoo.org>