Comment out sd_is_mq()
sd_is_mq() calls mq_getattr, which requires -lrt on systems using
uclibc. Ordinarily, I would write an autotools check to handle this, but
we do not use this code for anything, so I am commenting it out until a
decision on the fate of this code is made.

Signed-off-by: Richard Yao <ryao@gentoo.org>
Signed-off-by: Anthony G. Basile <blueness@gentoo.org>

temp