Forward-port of libudev.so.0 code from libudev-171
Add --enable-legacylib option to configure, allowing for build and
installation of libudev.so.0 shared library for supporting pre-udev-183
software.  Library is installed to rootlibdir.

Note that this legacy library will not be maintained, and should not be used
for anything other than supporting a working system until all software
on the system is migrated to use libudev.so.1

Signed-off-by: Anthony G. Basile <blueness@gentoo.org>