udev: export tags of "dead" device nodes to /run/udev/static_node-tags/
This is upstream 84b6ad702e64db534f67ce32d4dd2fec00a16784

Based on a patch by Kay Sievers.

A tag is exported at boot as a symlinks to the device node in the folder
/run/udev/static_node-tags/<tagname>/, if the device node exists.

These tags are cleaned up by udevadm info --cleanup-db, but are otherwise
never removed.

Signed-off-by: Anthony G. Basile <blueness@gentoo.org>