src/libudev/{strv,util}.{h,c}: reorder functions
We reorder the functions prototypes and definitions to match
upstream as in commit aa417a4d83999f6d7f092161d5c411b8cbce9977.
The order was lost when that commit was revert and the functions
re-introduced in later commits.

Preserving the order helps to better track upstream changes when
doing a diff between files.

Signed-off-by: Anthony G. Basile <blueness@gentoo.org>