configure.ac: add option to disable building programs
In some cases (e.g. trimmed-down chroot or containers), libudev is the
only thing needed, with udevd and assorted programs totally useless
(e.g. because /dev is bind-mounted from the real one and managed
out-side the chroot/container).

Add an option to ./configure to enable/disable building the programs;
this option defaults to "enable", so that it is backward compatible with
existing build procedure, and because it by default makes sense to have
udevd et al.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Anthony G. Basile <blueness@gentoo.org>