Remove references to /run/udev/hwdb.d
This directory never seemed to exist. The wording is very similar to
that of the Rules Files section, so it is possible that the same text
was copied and pasted by mistake. The HWDB never looked for files in
/run.

Having a non-volatile index of volatile files would probably not be
useful.