rules: Enable media-presence polling for all removable devices.
The existing rule would only be effective when the block drivers are compiled
as modules as opposed to being builtin.  On some systems such as Guix System,
the 'sr_mod' driver for optical discs is builtin the kernel instead of a
module, and the absence of polling means there are no uevents emitted for
udisks to act on.

This change was originally contributed to Debian by Martin Pitt to fix
<https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=713877>, later removed as it
was deemed no longer needed.  It appears still useful on eudev and elogind
systems like Guix System and others.

* rules.d/60-block.rules: Apply kernel polling to all removable devices instead
of just those of the block category handled by a kernel module.